pqms_backend/requirements.txt

Django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-filter>=24.3
drf-spectacular>=0.27
Pillow>=11.0
